+++ main/boards/nerdqaxeplus.h
@@ -30,7 +30,7 @@ class NerdQaxePlus : public Board {
     virtual bool setVoltage(float core_voltage);
 
     virtual void setFanPolarity(bool invert);
-    virtual void setFanSpeedCh(int channel, float perc);
+    virtual void setFanSpeedCh(int channel, int perc);
     virtual void getFanSpeedCh(int channel, uint16_t *rpm);
 
     virtual float getTemperature(int index);
